Rupali: Updated Admin Panel

+++ controller/removeUserByAdmin.js
@@ -11,19 +11,33 @@ function removeUserByAdmin(req, res) {
       .json({ error: "user_id (mobile_number) is required" });
   }
 
-  const deleteQuery = `DELETE FROM users WHERE user_id = ?`;
+  // first delete address
+  const deleteAddressQuery = `DELETE FROM addresses WHERE user_id = ?`;
 
-  connection.query(deleteQuery, [user_id], (err, result) => {
-    if (err) {
-      console.error("Error deleting user:", err);
-      return res.status(500).json({ error: "Error deleting user" });
+  connection.query(deleteAddressQuery, [user_id], (addressErr) => {
+    if (addressErr) {
+      console.error("Error deleting address:", addressErr);
+      return res.status(500).json({ error: "Error deleting address" });
     }
 
-    if (result.affectedRows === 0) {
-      return res.status(404).json({ error: "User not found" });
-    }
+    // then delete user
+    const deleteUserQuery = `DELETE FROM users WHERE user_id = ?`;
 
-    return res.status(200).json({ message: "User deleted successfully" });
+    connection.query(deleteUserQuery, [user_id], (userErr, result) => {
+      if (userErr) {
+        console.error("Error deleting user:", userErr);
+        return res.status(500).json({ error: "Error deleting user" });
+      }
+
+      if (result.affectedRows === 0) {
+        return res.status(404).json({ error: "User not found" });
+      }
+
+      return res.status(200).json({
+        success: true,
+        message: "User and address deleted successfully",
+      });
+    });
   });
 }
